Relocated SearchByCaloriesAndProteinUseCase and GetFastHealthyMealsUseCase to org.example.logic.usecase package to improve code structure and maintainability
